+++ src/main/java/seedu/delino/MainApp.java
@@ -24,14 +24,15 @@ import seedu.delino.model.ReadOnlyUserPrefs;
 import seedu.delino.model.ReturnOrderBook;
 import seedu.delino.model.UserPrefs;
 import seedu.delino.model.util.SampleDataUtil;
-import seedu.delino.storage.JsonOrderBookStorage;
-import seedu.delino.storage.JsonReturnOrderBookStorage;
 import seedu.delino.storage.JsonUserPrefsStorage;
-import seedu.delino.storage.OrderBookStorage;
-import seedu.delino.storage.ReturnOrderBookStorage;
 import seedu.delino.storage.Storage;
 import seedu.delino.storage.StorageManager;
 import seedu.delino.storage.UserPrefsStorage;
+import seedu.delino.storage.order.JsonOrderBookStorage;
+import seedu.delino.storage.order.OrderBookStorage;
+import seedu.delino.storage.returnorder.JsonReturnOrderBookStorage;
+import seedu.delino.storage.returnorder.ReturnOrderBookStorage;
+
 import seedu.delino.ui.Ui;
 import seedu.delino.ui.UiManager;
