+++ src/main/java/seedu/delino/storage/StorageManager.java
@@ -11,6 +11,8 @@ import seedu.delino.model.ReadOnlyOrderBook;
 import seedu.delino.model.ReadOnlyReturnOrderBook;
 import seedu.delino.model.ReadOnlyUserPrefs;
 import seedu.delino.model.UserPrefs;
+import seedu.delino.storage.order.OrderBookStorage;
+import seedu.delino.storage.returnorder.ReturnOrderBookStorage;
 
 /**
  * Manages storage of OrderBook and ReturnOrderBook data in local storage.
