+++ src/main/java/seedu/delino/storage/order/JsonSerializableDeliveryOrderBook.java
@@ -1,4 +1,4 @@
-package seedu.delino.storage;
+package seedu.delino.storage.order;
 
 import java.util.ArrayList;
 import java.util.List;
