+++ src/test/java/seedu/delino/storage/order/JsonSerializableDeliveryOrderBookTest.java
@@ -1,4 +1,4 @@
-package seedu.delino.storage;
+package seedu.delino.storage.order;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static seedu.delino.testutil.Assert.assertThrows;
